requirements.txt
langchain==0.3.19
langchain-community==0.3.18
langchain-openai==0.3.7
langchain-core==0.3.40
langchain-aws==0.2.1
boto3==1.34.144
sqlalchemy==2.0.28
python-dotenv==1.0.1
fastapi==0.115.8
psycopg2-binary==2.9.9
pgvector==0.2.5
uvicorn==0.28.0
pypdf==4.1.0
unstructured==0.16.11
markdown==3.6
networkx==3.2.1
pandas==2.2.1
openpyxl==3.1.2
docx2txt==0.8
pypandoc==1.13
PyJWT==2.8.0
asyncpg==0.29.0
python-multipart==0.0.19
sentence_transformers==3.1.1
aiofiles==23.2.1
rapidocr-onnxruntime==1.3.24
opencv-python-headless==4.9.0.80
pymongo==4.6.3
langsmith==0.3.11
langchain-mongodb==0.5.0
langchain-ollama==0.2.0
langchain-huggingface==0.1.0
cryptography==43.0.1
python-magic==0.4.27
python-pptx==0.6.23
xlrd==2.0.1
pytest==8.3.4